requirements.txt
requests==2.31.0
python-dotenv==1.0.0
dramatiq
dramatiq[rabbitmq]
apscheduler
pika
redis
mysql-connector-python
minio